Add fallback for external functions
based on:
https://github.com/inkle/ink/blob/master/Documentation/RunningYourInk.md#fallbacks-for-external-functions

it is no possible to implement a dummey/error function for each external
functions. It must have ben the same name and same amount of arguments
and will be executed if the external functio was not bind when needed.
IF the external functions will get bind later, then it will be called
the next time.

Introduces new flag: "FALLBACK_FUNCTION" to add a new execution
dimension
Introduce new value type: `ex_fn_not_found` to signal failed excetrnal
executio